uhsnc

+++ Seminar_05/view.py
@@ -2,7 +2,7 @@ def view_data(data, title):
     print(f'{title} = {data}')
 
 def get_value():
-    return int(input('value = '))
+    return float(input('value = '))
 
 def get_function():
     return str(input('input on of: +, -, *, / -->> '))
